refactor(solver): split conditional evaluation tests
